+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - programs(treasury): Corrected data error in GT buyback message.
+- sdk(solana-utils): Fixed missing payer signer in transaction size calculation.
 
 ## [0.5.0] - 2025-05-16